zkService/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.0
pydantic-settings==2.3.4
pyzk==0.9
websockets==12.0
requests==2.32.3